Split up Document field into better organised files
