+++ packages/fields-document/src/DocumentEditor/documentFeatures.js
@@ -1,0 +1,7 @@
+import { createContext, useContext } from 'react';
+
+export const DocumentFeaturesContext = createContext({});
+
+export const useDocumentFeatures = () => {
+  return useContext(DocumentFeaturesContext);
+};
