+++ packages/fields-document/src/views/Field.js
@@ -1,39 +1,23 @@
 /** @jsx jsx */
 
 import { jsx } from '@emotion/core';
-import { createContext, useCallback, useMemo, useContext } from 'react';
-import { Slate, Editable, ReactEditor, useSlate, withReact } from 'slate-react';
-import { createEditor, Editor, Transforms, Text, Element, Path, Node, Range } from 'slate';
+import { useCallback, useMemo } from 'react';
+import { Slate, Editable, useSlate, withReact } from 'slate-react';
+import { createEditor, Editor, Transforms, Text } from 'slate';
 import { withHistory } from 'slate-history';
 
 import { FieldContainer, FieldLabel, FieldDescription } from '@arch-ui/fields';
 
-const DocumentFeaturesContext = createContext({});
-const useDocumentFeatures = () => {
-  return useContext(DocumentFeaturesContext);
-};
-
-const Button = ({ isDisabled, isPressed, ...props }) => (
-  <button
-    type="button"
-    disabled={isDisabled}
-    css={{
-      background: isPressed ? '#f3f3f3' : 'white',
-      borderColor: isDisabled ? '#eee' : isPressed ? '#aaa' : '#ccc',
-      borderStyle: 'solid',
-      borderWidth: 1,
-      borderRadius: 5,
-      boxShadow: isPressed ? 'inset 0px 3px 5px -4px rgba(0,0,0,0.50)' : undefined,
-      color: isDisabled ? '#999' : '#172B4D',
-      marginRight: 4,
-      pointerEvents: isDisabled ? 'none' : undefined,
-      ':hover': {
-        borderColor: isPressed ? '#666' : '#999',
-      },
-    }}
-    {...props}
-  />
-);
+import {
+  AccessBoundaryElement,
+  insertAccessBoundary,
+  isInsideAccessBoundary,
+  withAccess,
+} from '../DocumentEditor/access';
+import { Button } from '../DocumentEditor/components';
+import { DocumentFeaturesContext } from '../DocumentEditor/documentFeatures';
+import { withParagraphs } from '../DocumentEditor/paragraphs';
+import { isBlockActive } from '../DocumentEditor/utils';
 
 const DocumentEditor = {
   // Bold
@@ -84,14 +68,6 @@ const DocumentEditor = {
       { match: n => Editor.isBlock(editor, n) }
     );
   },
-};
-
-const isBlockActive = (editor, format) => {
-  const [match] = Editor.nodes(editor, {
-    match: n => n.type === format,
-  });
-
-  return !!match;
 };
 
 const getKeyDownHandler = editor => event => {
@@ -170,168 +146,6 @@ const Toolbar = () => {
       >
         + Access Boundary
       </Button>
-    </div>
-  );
-};
-
-/* Paragraphs */
-
-const withParagraphs = editor => {
-  const { normalizeNode } = editor;
-
-  editor.normalizeNode = entry => {
-    const [node, path] = entry;
-
-    // If the element is a paragraph, ensure its children are valid.
-    if (Element.isElement(node)) {
-      if (!node.type) {
-        Transforms.setNodes(editor, { type: 'paragraph' }, { at: path });
-        return;
-      }
-      if (!node.type || node.type === 'paragraph') {
-        for (const [child, childPath] of Node.children(editor, path)) {
-          if (Element.isElement(child) && !editor.isInline(child)) {
-            Transforms.unwrapNodes(editor, { at: childPath });
-            return;
-          }
-        }
-      }
-    }
-
-    // Fall back to the original `normalizeNode` to enforce other constraints.
-    normalizeNode(entry);
-  };
-
-  return editor;
-};
-
-/* Access Control */
-
-const ROLES = [
-  { label: 'Public', value: 'public' },
-  { label: 'Members', value: 'member' },
-  { label: 'Admins', value: 'admin' },
-];
-
-const getBlockAboveSelection = editor =>
-  Editor.above(editor, {
-    match: n => Editor.isBlock(editor, n),
-  }) || [editor, []];
-
-const isBlockTextEmpty = node => {
-  const lastChild = node.children[node.children.length - 1];
-  return Text.isText(lastChild) && !lastChild.text.length;
-};
-
-// Access Boundary
-const isInsideAccessBoundary = editor => {
-  return isBlockActive(editor, 'access-boundary');
-};
-const insertAccessBoundary = editor => {
-  if (isBlockActive(editor, 'access-boundary')) return;
-  const { selection } = editor;
-  const isCollapsed = selection && Range.isCollapsed(selection);
-  const [block] = getBlockAboveSelection(editor);
-  if (!!block && isCollapsed && isBlockTextEmpty(block)) {
-    const element = { type: 'access-boundary', roles: [], children: [] };
-    Transforms.wrapNodes(editor, element);
-  } else {
-    const children = [{ type: 'paragraph', children: [{ text: '' }] }];
-    const element = { type: 'access-boundary', roles: [], children };
-    Transforms.insertNodes(editor, element, { select: true });
-  }
-};
-
-const withAccess = editor => {
-  const { insertBreak } = editor;
-  editor.insertBreak = () => {
-    const [block] = getBlockAboveSelection(editor);
-
-    if (block && isBlockTextEmpty(block)) {
-      const accessBoundary = Editor.above(editor, {
-        match: n => n.type === 'access-boundary',
-      });
-
-      if (accessBoundary) {
-        const [, path] = accessBoundary;
-        Transforms.insertNodes(
-          editor,
-          { type: 'paragraph', children: [{ text: '' }] },
-          {
-            at: Path.next(path),
-            select: true,
-          }
-        );
-        return;
-      }
-    }
-
-    insertBreak();
-  };
-  return editor;
-};
-
-const AccessBoundaryElement = ({ attributes, children, element }) => {
-  const editor = useSlate();
-  const { access } = useDocumentFeatures();
-  const elementRoles = Array.isArray(element.roles) ? element.roles : [];
-  // TODO: handle case where no access or roles are defined
-  // TODO: validate elementRoles against defined access roles
-  return (
-    <div
-      css={{
-        margin: '8px 0',
-        border: '3px dashed #eee',
-        borderRadius: 5,
-      }}
-      {...attributes}
-    >
-      <div
-        contentEditable={false}
-        style={{
-          backgroundColor: '#f9f9f9',
-          borderBottom: '1px solid #eee',
-          padding: 8,
-          fontSize: 14,
-          color: '#d64242',
-          fontWeight: 600,
-          userSelect: 'none',
-        }}
-      >
-        <span css={{ marginRight: 8 }}>Restrict to:</span>
-        {access.roles.map(role => {
-          const roleIsSelected = elementRoles.includes(role.value);
-          return (
-            <Button
-              key={role.value}
-              isPressed={roleIsSelected}
-              onMouseDown={event => {
-                event.preventDefault();
-                const path = ReactEditor.findPath(editor, element);
-                const newRoles = roleIsSelected
-                  ? elementRoles.filter(i => i !== role.value)
-                  : [...elementRoles, role.value];
-                Transforms.setNodes(editor, { roles: newRoles }, { at: path });
-              }}
-            >
-              {role.label}
-            </Button>
-          );
-        })}
-        <Button
-          css={{ float: 'right' }}
-          onMouseDown={event => {
-            event.preventDefault();
-            const path = ReactEditor.findPath(editor, element);
-            Transforms.unwrapNodes(editor, { at: path });
-            Transforms.select(editor, path);
-            Transforms.collapse(editor, { edge: 'start' });
-          }}
-        >
-          Remove
-        </Button>
-      </div>
-      <div css={{ margin: 8 }}>{children}</div>
     </div>
   );
 };
